+++ nomos-libp2p/src/behaviour/kademlia/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod behaviour_ext;
+pub mod swarm_ext;
