+++ nomos-libp2p/src/lib.rs
@@ -1,14 +1,9 @@
-mod behaviour;
+pub mod behaviour;
 mod config;
 pub mod protocol_name;
 mod swarm;
 
-use blake2::{
-    digest::{consts::U32, Digest as _},
-    Blake2b,
-};
 pub use config::{secret_key_serde, IdentifySettings, KademliaSettings, SwarmConfig};
-use libp2p::gossipsub::{Message, MessageId};
 pub use libp2p::{
     self,
     core::upgrade,
@@ -19,10 +14,4 @@ pub use libp2p::{
 };
 pub use multiaddr::{multiaddr, Multiaddr, Protocol};
 
-pub use crate::{behaviour::*, swarm::*};
-
-fn compute_message_id(message: &Message) -> MessageId {
-    let mut hasher = Blake2b::<U32>::new();
-    hasher.update(&message.data);
-    MessageId::from(hasher.finalize().to_vec())
-}
+pub use crate::swarm::*;
