+++ nomos-services/network/src/backends/libp2p/swarm/identify.rs
@@ -1,3 +1,5 @@
+use std::collections::HashSet;
+
 use nomos_libp2p::libp2p::identify;
 
 use crate::backends::libp2p::swarm::SwarmHandler;
@@ -11,19 +13,19 @@ impl SwarmHandler {
                     peer_id,
                     info.listen_addrs
                 );
-                let kad_protocol_names = self.swarm.get_kademlia_protocol_names();
+                let kad_protocol_names = self
+                    .swarm
+                    .get_kademlia_protocol_names()
+                    .collect::<HashSet<_>>();
                 if info
                     .protocols
                     .iter()
-                    .any(|p| kad_protocol_names.contains(&p.to_string()))
+                    .any(|p| kad_protocol_names.contains(&p))
                 {
                     // we need to add the peer to the kademlia routing table
                     // in order to enable peer discovery
                     for addr in &info.listen_addrs {
-                        self.swarm
-                            .swarm_mut()
-                            .behaviour_mut()
-                            .kademlia_add_address(peer_id, addr.clone());
+                        self.swarm.kademlia_add_address(peer_id, addr.clone());
                     }
                 }
             }
